# sirius_rest_service/openapi-spec.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SIRIUS Nightsky API",
    "description": "REST API that provides the full functionality of SIRIUS and its web services as background service. It is intended as entry-point for scripting languages and software integration SDKs.This API is exposed by SIRIUS 6.0.0-SNAPSHOT",
    "version": "2.1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/api/projects/{projectId}/compounds": {
      "get": {
        "tags": [
          "Compounds"
        ],
        "summary": "List of all available compounds (group of ion identities) in the given project-space.",
        "description": "List of all available compounds (group of ion identities) in the given project-space.",
        "operationId": "getCompounds",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to read from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CompoundOptField"
              },
              "default": []
            }
          },
          {
            "name": "optFieldsFeatures",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AlignedFeatureOptField"
              },
              "default": []
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compounds with additional optional fields (if specified).",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Compound"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Compounds"
        ],
        "summary": "Import Compounds and its contained features.",
        "description": "Import Compounds and its contained features. Compounds and Features must not exist in the project.\n Otherwise, they will exist twice.",
        "operationId": "addCompounds",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to import into.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CompoundOptField"
              },
              "default": []
            }
          },
          {
            "name": "optFieldsFeatures",
            "in": "query",
            "description": "set of optional fields of the nested features to be included. Use \u0027none\u0027 to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AlignedFeatureOptField"
              },
              "default": []
            }
          }
        ],
        "requestBody": {
          "description": "the compound data to be imported",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/CompoundImport"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "the Compounds that have been imported with specified optional fields",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Compound"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/projects/{projectId}/compounds/{compoundId}": {
      "get": {
        "tags": [
          "Compounds"
        ],
        "summary": "Get compound (group of ion identities) with the given identifier from the specified project-space.",
        "description": "Get compound (group of ion identities) with the given identifier from the specified project-space.",
        "operationId": "getCompound",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to read from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compoundId",
            "in": "path",
            "description": "identifier of the compound (group of ion identities) to access.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CompoundOptField"
              },
              "default": []
            }
          },
          {
            "name": "optFieldsFeatures",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AlignedFeatureOptField"
              },
              "default": []
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compounds with additional optional fields (if specified).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Compound"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Compounds"
        ],
        "summary": "Delete compound (group of ion identities) with the given identifier (and the included features) from the\n specified project-space.",
        "description": "Delete compound (group of ion identities) with the given identifier (and the included features) from the\n specified project-space.",
        "operationId": "deleteCompound",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to delete from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "compoundId",
            "in": "path",
            "description": "identifier of the compound to delete.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/{projectId}/compounds/page": {
      "get": {
        "tags": [
          "Compounds"
        ],
        "summary": "Page of available compounds (group of ion identities) in the given project-space.",
        "description": "Page of available compounds (group of ion identities) in the given project-space.",
        "operationId": "getCompoundsPaged",
        "parameters": [
          {
            "name": "projectId",
            "in": "path",
            "description": "project-space to read from.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Zero-based page index (0..N)",
            "required": false,
            "schema": {
              "minimum": 0,
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "The size of the page to be returned",
            "required": false,
            "schema": {
              "minimum": 1,
              "type": "integer",
              "default": 20
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sorting criteria in the format: property,(asc|desc). Default sort order is ascending. Multiple sort criteria are supported.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          },
          {
            "name": "optFields",
            "in": "query",
            "description": "set of optional fields to be included. Use \u0027none\u0027 only to override defaults.",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CompoundOptField"
              },
              "default": []
            }
          },
          {
            "name": "optFieldsFeatures",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/AlignedFeatureOptField"
              },
              "default": []
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compounds with additional optional fields (if specified).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageCompound"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BasicSpectrum": {
        "required": [
          "peaks"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Optional Displayable name of this spectrum.",
            "nullable": true
          },
          "msLevel": {
            "type": "integer",
            "description": "MS level of the measured spectrum.\n Artificial spectra with no msLevel (e.g. Simulated Isotope patterns) use null or zero",
            "format": "int32",
            "nullable": true
          },
          "collisionEnergy": {
            "type": "string",
            "description": "Collision energy used for MS/MS spectra\n Null for spectra where collision energy is not applicable",
            "nullable": true
          },
          "precursorMz": {
            "type": "number",
            "description": "Precursor m/z of the MS/MS spectrum\n Null for spectra where precursor m/z is not applicable",
            "format": "double",
            "nullable": true
          },
          "scanNumber": {
            "type": "integer",
            "description": "Scan number of the spectrum.\n Might be null for artificial spectra with no scan number (e.g. Simulated Isotope patterns or merged spectra)",
            "format": "int32",
            "nullable": true
          },
          "peaks": {
            "type": "array",
            "description": "The peaks of this spectrum which might contain additional annotations such as molecular formulas.",
            "items": {
              "$ref": "#/components/schemas/SimplePeak"
            }
          },
          "absIntensityFactor": {
            "type": "number",
            "description": "Factor to convert relative intensities to absolute intensities.\n Might be null or 1 for spectra where absolute intensities are not available (E.g. artificial or merged spectra)",
            "format": "double",
            "nullable": true
          }
        }
      },
      "CompoundImport": {
        "required": [
          "features"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Some (optional) human-readable name",
            "nullable": true
          },
          "features": {
            "type": "array",
            "description": "The features this compound consists of.",
            "items": {
              "$ref": "#/components/schemas/FeatureImport"
            }
          }
        }
      },
      "FeatureImport": {
        "required": [
          "adduct",
          "ionMass",
          "ms1Spectra",
          "ms2Spectra"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "featureId": {
            "type": "string",
            "nullable": true
          },
          "ionMass": {
            "type": "number",
            "format": "double"
          },
          "adduct": {
            "type": "string",
            "description": "Adduct of this feature. If not know specify [M+?]+ or [M+?]- to define the charge"
          },
          "rtStartSeconds": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "rtEndSeconds": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "mergedMs1": {
            "$ref": "#/components/schemas/BasicSpectrum"
          },
          "ms1Spectra": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicSpectrum"
            }
          },
          "ms2Spectra": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicSpectrum"
            }
          }
        }
      },
      "SimplePeak": {
        "type": "object",
        "properties": {
          "mz": {
            "type": "number",
            "format": "double"
          },
          "intensity": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "CompoundOptField": {
        "type": "string",
        "nullable": true,
        "enum": [
          "none",
          "consensusAnnotations",
          "consensusAnnotationsDeNovo",
          "customAnnotations"
        ]
      },
      "AlignedFeatureOptField": {
        "type": "string",
        "nullable": true,
        "enum": [
          "none",
          "msData",
          "topAnnotations",
          "topAnnotationsDeNovo"
        ]
      },
      "AlignedFeature": {
        "type": "object",
        "properties": {
          "alignedFeatureId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "index": {
            "type": "integer",
            "format": "int64"
          },
          "ionMass": {
            "type": "number",
            "format": "double"
          },
          "adduct": {
            "type": "string"
          },
          "rtStartSeconds": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "rtEndSeconds": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "msData": {
            "$ref": "#/components/schemas/MsData"
          },
          "topAnnotations": {
            "$ref": "#/components/schemas/FeatureAnnotations"
          },
          "topAnnotationsDeNovo": {
            "$ref": "#/components/schemas/FeatureAnnotations"
          },
          "computing": {
            "type": "boolean",
            "description": "Write lock for this feature. If the feature is locked no write operations are possible.\n True if any computation is modifying this feature or its results"
          }
        },
        "description": "The AlignedFeature contains the ID of a featured (aligned over runs) together with some read-only information\n that might be displayed in some summary view."
      },
...
      "Compound": {
        "type": "object",
        "properties": {
          "compoundId": {
            "type": "string",
            "description": "uid of this compound Entity"
          },
          "name": {
            "type": "string",
            "description": "Some (optional) human-readable name",
            "nullable": true
          },
          "rtStartSeconds": {
            "type": "number",
            "description": "The merged/consensus retention time start (earliest rt) of this compound",
            "format": "double",
            "nullable": true
          },
          "rtEndSeconds": {
            "type": "number",
            "description": "The merged/consensus retention time end (latest rt) of this compound",
            "format": "double",
            "nullable": true
          },
          "neutralMass": {
            "type": "number",
            "description": "Neutral mass of this compound. Ion masse minus the mass of the assigned adduct of each feature of\n this compound should result in the same neutral mass",
            "format": "double",
            "nullable": true
          },
          "features": {
            "type": "array",
            "description": "List of aligned features (adducts) that belong to the same (this) compound",
            "items": {
              "$ref": "#/components/schemas/AlignedFeature"
            }
          },
          "consensusAnnotations": {
            "$ref": "#/components/schemas/ConsensusAnnotationsCSI"
          },
          "consensusAnnotationsDeNovo": {
            "$ref": "#/components/schemas/ConsensusAnnotationsDeNovo"
          },
          "customAnnotations": {
            "$ref": "#/components/schemas/ConsensusAnnotationsCSI"
          }
        }
      },
...
      "CompoundClasses": {
        "type": "object",
        "properties": {
          "npcPathway": {
            "$ref": "#/components/schemas/CompoundClass"
          },
          "npcSuperclass": {
            "$ref": "#/components/schemas/CompoundClass"
          },
          "npcClass": {
            "$ref": "#/components/schemas/CompoundClass"
          },
          "classyFireLineage": {
            "type": "array",
            "description": "Most likely ClassyFire lineage from ordered from least specific to most specific class\n classyFireLineage.get(classyFireLineage.size() - 1) gives the most specific ClassyFire compound class annotation",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CompoundClass"
            }
          },
          "classyFireAlternatives": {
            "type": "array",
            "description": "Alternative ClassyFire classes with high probability that do not fit into the linage",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/CompoundClass"
            }
          }
        },
        "description": "Container class that holds the most likely compound class for different levels of each ontology for a\n certain Compound/Feature/FormulaCandidate/PredictedFingerprint.",
        "nullable": true
      },
      "ConsensusAnnotationsCSI": {
        "type": "object",
        "properties": {
          "molecularFormula": {
            "type": "string",
            "description": "Molecular formula of the consensus annotation\n Might be null if no consensus formula is available.",
            "nullable": true
          },
          "compoundClasses": {
            "$ref": "#/components/schemas/CompoundClasses"
          },
          "supportingFeatureIds": {
            "type": "array",
            "description": "FeatureIds where the topAnnotation supports this annotation.",
            "nullable": true,
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "selectionCriterion": {
            "$ref": "#/components/schemas/ConsensusCriterionCSI"
          },
          "csiFingerIdStructure": {
            "$ref": "#/components/schemas/StructureCandidate"
          },
          "confidenceExactMatch": {
            "type": "number",
            "description": "Confidence value that represents the certainty that reported consensus structure is exactly the measured one\n If multiple features support this consensus structure the maximum confidence is reported",
            "format": "double",
            "nullable": true
          },
          "confidenceApproxMatch": {
            "type": "number",
            "description": "Confidence value that represents the certainty that the exact consensus structure or a very similar\n structure (e.g. measured by Maximum Common Edge Subgraph Distance) is the measured one.\n If multiple features support this consensus structure the maximum confidence is reported",
            "format": "double",
            "nullable": true
          }
        },
        "nullable": true
      },
      "ConsensusAnnotationsDeNovo": {
        "type": "object",
        "properties": {
          "molecularFormula": {
            "type": "string",
            "description": "Molecular formula of the consensus annotation\n Might be null if no consensus formula is available.",
            "nullable": true
          },
          "compoundClasses": {
            "$ref": "#/components/schemas/CompoundClasses"
          },
          "supportingFeatureIds": {
            "type": "array",
            "description": "FeatureIds where the topAnnotation supports this annotation.",
            "nullable": true,
            "items": {
              "type": "string",
              "nullable": true
            }
          },
          "selectionCriterion": {
            "$ref": "#/components/schemas/ConsensusCriterionDeNovo"
          }
        },
        "nullable": true
      },
      "ConsensusCriterionCSI": {
        "type": "string",
        "nullable": true,
        "enum": [
          "MAJORITY_STRUCTURE",
          "CONFIDENCE_STRUCTURE",
          "SINGLETON_STRUCTURE",
          "MAJORITY_FORMULA",
          "TOP_FORMULA",
          "SINGLETON_FORMULA"
        ]
      },
      "ConsensusCriterionDeNovo": {
        "type": "string",
        "nullable": true,
        "enum": [
          "MAJORITY_FORMULA",
          "TOP_FORMULA",
          "SINGLETON_FORMULA"
        ]
      },
      "DBLink": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "id": {
            "type": "string"
          }
        },
        "nullable": true
      },
...
      "FeatureAnnotations": {
        "type": "object",
        "properties": {
          "formulaAnnotation": {
            "$ref": "#/components/schemas/FormulaCandidate"
          },
          "structureAnnotation": {
            "$ref": "#/components/schemas/StructureCandidateScored"
          },
          "compoundClassAnnotation": {
            "$ref": "#/components/schemas/CompoundClasses"
          }
        },
        "description": "Summary of the results of a feature (aligned over runs). Can be added to a AlignedFeature.\n The different annotation fields within this summary object are null if the corresponding\n feature does not contain the represented results. If fields are non-null\n the corresponding result has been computed but might still be empty.",
        "nullable": true
      },
...
      "MsData": {
        "required": [
          "ms1Spectra",
          "ms2Spectra"
        ],
        "type": "object",
        "properties": {
          "mergedMs1": {
            "$ref": "#/components/schemas/BasicSpectrum"
          },
          "mergedMs2": {
            "$ref": "#/components/schemas/BasicSpectrum"
          },
          "ms1Spectra": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicSpectrum"
            }
          },
          "ms2Spectra": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BasicSpectrum"
            }
          }
        },
        "description": "The MsData wraps all spectral input data belonging to a feature.\n \u003cp\u003e\n Each Feature has:\n - One merged MS/MS spectrum (optional)\n - One merged MS spectrum (optional)\n - many MS/MS spectra\n - many MS spectra\n \u003cp\u003e\n Each non-merged spectrum has an index which can be used to access the spectrum.\n \u003cp\u003e\n In the future we might add some additional information like chromatographic peak or something similar",
        "nullable": true
      },
...
      "StructureCandidate": {
        "type": "object",
        "properties": {
          "inchiKey": {
            "type": "string"
          },
          "smiles": {
            "type": "string"
          },
          "structureName": {
            "type": "string",
            "nullable": true
          },
          "xlogP": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "dbLinks": {
            "type": "array",
            "description": "List of structure database links belonging to this structure candidate\n OPTIONAL: needs to be added by parameter",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/DBLink"
            }
          },
          "spectralLibraryMatches": {
            "type": "array",
            "description": "List of spectral library matches belonging to this structure candidate\n OPTIONAL: needs to be added by parameter",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/SpectralLibraryMatch"
            }
          }
        },
        "nullable": true
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "paged": {
            "type": "boolean"
          },
          "unpaged": {
            "type": "boolean"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          }
        }
      },
      "PageCompound": {
        "type": "object",
        "properties": {
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "last": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "first": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Compound"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}